Move queue module into orchestration directory
